Fixes to LAStoPAF, gap improvement, refactoring of k-mer and post tables
A bug in the -x option of LAStoPAF wherein the last alignment segment was M's and not X's and ='s is fixed.

The new gap improver module was not handling boundary extension correctly which was causing memory overrights (but not a crash).

All IO operations are now checked and Gindex is guaranteed not to leave any detritus if it crashes.

Gindex now produces only T part files as opposed to T^2.