Implementation of soft masks, improved ALNto(PAF|PSL) & GIXmake, -L option added
Main addition is recognition of soft-masks in .fasta files using upper/lower case letters to denote said.  The masks are recorded in the upgraded GDB and in a form suitable for the index merge in GIX indices.  This necessitated a new form of GIX and so FastGA had to be significantly updated.  The old GIX's are still recognized and supported.

Formatted IO was slow in ALNtoPAF and ALNtoPSL and has been replaced with direct character output resulting in substantial speed increases.

GIXmake has been substantially improved to use less memory and with a better threading strategy is about 50% faster than the last version.

For HPC runs on a cluster, a -L for generating a log file has been added.